+++ glean-core/src/lib.rs
@@ -161,7 +161,7 @@ pub struct Configuration {
 ///
 /// call_counter.add(&glean, 1);
 ///
-/// glean.submit_ping(&ping, None).unwrap();
+/// glean.submit_ping(&ping, None);
 /// ```
 ///
 /// ## Note
@@ -454,8 +454,8 @@ impl Glean {
         } else {
             Some("set_upload_enabled")
         };
-        if let Err(err) = self.internal_pings.deletion_request.submit(self, reason) {
-            log::error!("Failed to submit deletion-request ping on optout: {}", err);
+        if !self.internal_pings.deletion_request.submit(self, reason) {
+            log::error!("Failed to submit deletion-request ping on optout.");
         }
         self.clear_metrics();
         self.upload_enabled = false;
@@ -626,10 +626,10 @@ impl Glean {
     /// # Returns
     ///
     /// Whether the ping was succesfully assembled and queued.
-    pub fn submit_ping(&self, ping: &PingType, reason: Option<&str>) -> Result<bool> {
+    pub fn submit_ping(&self, ping: &PingType, reason: Option<&str>) -> bool {
         if !self.is_upload_enabled() {
             log::info!("Glean disabled: not submitting any pings.");
-            return Ok(false);
+            return false;
         }
 
         let ping_maker = PingMaker::new();
@@ -641,7 +641,7 @@ impl Glean {
                     "No content for ping '{}', therefore no ping queued.",
                     ping.name
                 );
-                Ok(false)
+                false
             }
             Some(ping) => {
                 // This metric is recorded *after* the ping is collected (since
@@ -656,7 +656,12 @@ impl Glean {
                 if let Err(e) = ping_maker.store_ping(&self.get_data_path(), &ping) {
                     log::warn!("IO error while writing ping to file: {}. Enqueuing upload of what we have in memory.", e);
                     self.additional_metrics.io_errors.add(self, 1);
-                    let content = ::serde_json::to_string(&ping.content)?;
+                    // `serde_json::to_string` only fails if serialization of the content
+                    // fails or it contains maps with non-string keys.
+                    // However `ping.content` is already a `JsonValue`,
+                    // so both scenarios should be impossible.
+                    let content =
+                        ::serde_json::to_string(&ping.content).expect("ping serialization failed");
                     self.upload_manager.enqueue_ping(
                         self,
                         ping.doc_id,
@@ -664,8 +669,7 @@ impl Glean {
                         &content,
                         Some(ping.headers),
                     );
-                    // Not actually 100% 'Ok'. bug 1704606
-                    return Ok(true);
+                    return true;
                 }
 
                 self.upload_manager.enqueue_ping_from_file(self, &doc_id);
@@ -674,7 +678,8 @@ impl Glean {
                     "The ping '{}' was submitted and will be sent as soon as possible",
                     ping.name
                 );
-                Ok(true)
+
+                true
             }
         }
     }
@@ -699,11 +704,11 @@ impl Glean {
     /// # Errors
     ///
     /// If collecting or writing the ping to disk failed.
-    pub fn submit_ping_by_name(&self, ping_name: &str, reason: Option<&str>) -> Result<bool> {
+    pub fn submit_ping_by_name(&self, ping_name: &str, reason: Option<&str>) -> bool {
         match self.get_ping_by_name(ping_name) {
             None => {
                 log::error!("Attempted to submit unknown ping '{}'", ping_name);
-                Ok(false)
+                false
             }
             Some(ping) => self.submit_ping(ping, reason),
         }
@@ -917,8 +922,8 @@ impl Glean {
     /// This functions generates a baseline ping with reason `active`
     /// and then sets the dirty bit.
     pub fn handle_client_active(&mut self) {
-        if let Err(err) = self.internal_pings.baseline.submit(self, Some("active")) {
-            log::warn!("Failed to submit baseline ping on active: {}", err);
+        if !self.internal_pings.baseline.submit(self, Some("active")) {
+            log::info!("baseline ping not submitted on active");
         }
 
         self.set_dirty_flag(true);
@@ -929,12 +934,12 @@ impl Glean {
     /// This functions generates a baseline and an events ping with reason
     /// `inactive` and then clears the dirty bit.
     pub fn handle_client_inactive(&mut self) {
-        if let Err(err) = self.internal_pings.baseline.submit(self, Some("inactive")) {
-            log::warn!("Failed to submit baseline ping on inactive: {}", err);
+        if !self.internal_pings.baseline.submit(self, Some("inactive")) {
+            log::info!("baseline ping not submitted on inactive");
         }
 
-        if let Err(err) = self.internal_pings.events.submit(self, Some("inactive")) {
-            log::warn!("Failed to submit events ping on inactive: {}", err);
+        if !self.internal_pings.events.submit(self, Some("inactive")) {
+            log::info!("events ping not submitted on inactive");
         }
 
         self.set_dirty_flag(false);